requirements.txt
flask
flask-mail
flask-pymongo
python-dotenv
flask-cors
Werkzeug
dnspython